# perf: 개인 공간 렌더링 최적화 (resolve #186, resolve #188)