--- a/client/src/MainPage/MainWorld.tsx
+++ b/client/src/MainPage/MainWorld.tsx
@@ -21,7 +21,7 @@ export default function MainWorld() {
       <Ground />
       <Environments />
       <Player />
-      {/* <DevTools speed={40} /> */}
+      <DevTools speed={40} />
     </Physics>
   );
 }