--- a/client/src/MainPage/mapObjects/Environments/RoadSign/index.tsx
+++ b/client/src/MainPage/mapObjects/Environments/RoadSign/index.tsx
@@ -64,64 +64,60 @@ export default function RoadSign(
   return (
     // @ts-ignore
     <animated.group ref={ref} {...props} {...springs} dispose={null}>
-      <group rotation={[-Math.PI / 2, 0, 0]}>
-        <group rotation={[Math.PI / 2, 0, 0]}>
-          <group position={[0.21, 4.91, 0.02]}>
-            <mesh castShadow receiveShadow geometry={nodes.Object_6.geometry} material={materials["WoodLight.001"]}>
-              <Html
-                position={[0.1, 0, -0.1]}
-                rotation={[0, Math.PI / 2, 0]}
-                transform={!showModal}
-                occlude
-                zIndexRange={[1, 10]}
+      <group position={[0.21, 4.91, 0.02]}>
+        <mesh receiveShadow geometry={nodes.Object_6.geometry} material={materials["WoodLight.001"]}>
+          <Html
+            position={[0.1, 0, -0.1]}
+            rotation={[0, Math.PI / 2, 0]}
+            transform={!showModal}
+            occlude
+            zIndexRange={[1, 10]}
+          >
+            <div
+              className="road-sign"
+              onPointerEnter={() => setMove(false)}
+              onPointerLeave={() => setMove(true)}
+              onClick={() => setShowModal(true)}
+            >
+              <div>
+                <span>{showModal ? "" : "모뉴먼트 갤러리"}</span>
+              </div>
+              <div>
+                <span>{showModal ? "" : "사용법 및 소개 클릭!"}</span>
+              </div>
+              <button
+                className="sign-button"
+                onClick={(e) => {
+                  e.stopPropagation();
+                  props.setShow(false);
+                }}
               >
-                <div
-                  className="road-sign"
-                  onPointerEnter={() => setMove(false)}
-                  onPointerLeave={() => setMove(true)}
-                  onClick={() => setShowModal(true)}
+                {!showModal && <img width={6} src={CloseIcon} alt="closeIcon" />}
+              </button>
+            </div>
+            <FullScreenModal
+              css={{ width: "70vw", height: "80vh", opacity: "0.9" }}
+              show={showModal}
+              setShow={setShowModal}
+            >
+              <div className="modal" onWheel={(e) => e.stopPropagation()}>
+                <RoadSignHtml />
+                <button
+                  className="sign-modal-close-button"
+                  onClick={() => {
+                    setShowModal(false);
+                    setMove(true);
+                  }}
                 >
-                  <div>
-                    <span>{showModal ? "" : "모뉴먼트 갤러리"}</span>
-                  </div>
-                  <div>
-                    <span>{showModal ? "" : "사용법 및 소개 클릭!"}</span>
-                  </div>
-                  <button
-                    className="sign-button"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      props.setShow(false);
-                    }}
-                  >
-                    {!showModal && <img width={6} src={CloseIcon} alt="closeIcon" />}
-                  </button>
-                </div>
-                <FullScreenModal
-                  css={{ width: "70vw", height: "80vh", opacity: "0.9" }}
-                  show={showModal}
-                  setShow={setShowModal}
-                >
-                  <div className="modal" onWheel={(e) => e.stopPropagation()}>
-                    <RoadSignHtml />
-                    <button
-                      className="sign-modal-close-button"
-                      onClick={() => {
-                        setShowModal(false);
-                        setMove(true);
-                      }}
-                    >
-                      <img width={15} src={CloseIcon} alt="closeIcon" />
-                    </button>
-                  </div>
-                </FullScreenModal>
-              </Html>
-            </mesh>
-          </group>
-          <group position={[0.27, 5.36, -0.01]} rotation={[0, 0, -Math.PI / 2]}>
-            <mesh castShadow receiveShadow geometry={nodes.Object_8.geometry} material={materials["iron.001"]} />
-          </group>
-        </group>
+                  <img width={15} src={CloseIcon} alt="closeIcon" />
+                </button>
+              </div>
+            </FullScreenModal>
+          </Html>
+        </mesh>
+      </group>
+      <group position={[0.27, 5.36, -0.01]} rotation={[0, 0, -Math.PI / 2]}>
+        <mesh receiveShadow geometry={nodes.Object_8.geometry} material={materials["iron.001"]} />
       </group>
     </animated.group>
   );