--- a/client/src/MainPage/mapObjects/Portal.tsx
+++ b/client/src/MainPage/mapObjects/Portal.tsx
@@ -39,10 +39,10 @@ function PortalMesh() {
   return (
     <group dispose={null}>
       <mesh castShadow receiveShadow geometry={nodes.Plane002_1.geometry} material={materials["Material.003"]} />
-      <mesh castShadow receiveShadow geometry={nodes.Plane002_2.geometry} material={materials["Material.002"]} />
+      <mesh receiveShadow geometry={nodes.Plane002_2.geometry} material={materials["Material.002"]} />
       <group position={[0, 0.92, 0]}>
-        <mesh castShadow receiveShadow geometry={nodes.Plane003_2.geometry} material={materials["Material.004"]} />
-        <mesh castShadow receiveShadow geometry={nodes.Plane003_3.geometry} material={materials["Material.003"]} />
+        <mesh receiveShadow geometry={nodes.Plane003_2.geometry} material={materials["Material.004"]} />
+        <mesh receiveShadow geometry={nodes.Plane003_3.geometry} material={materials["Material.003"]} />
       </group>
     </group>
   );