… client/create.html
… client/gallery.html
… client/index.html
… client/public/favicon.ico/manifest.json
… client/src/assets/images/expand.png
… client/src/assets/images/fullscreen.svg
… client/src/assets/images/github.svg
… client/src/assets/images/header-logo.svg
… client/src/assets/images/headerLogo.svg
… client/src/assets/images/pause.png
… client/src/assets/images/pause.svg
… client/src/assets/images/play.png
… client/src/assets/images/play.svg
… client/src/assets/images/screenshot.svg
… client/src/assets/images/splash-logo-1x.png
… client/src/assets/images/splash-logo-2x.png
… client/src/assets/images/viewfinder.png
… client/src/components/Footer/index.tsx
… client/src/components/Header/index.tsx
… client/src/components/Header/style.scss
… client/src/components/ThemeSelector/index.tsx
… client/src/hooks/useKeyMovement.ts
… client/src/utils/randomPosition.json
… server/index.js
… server/service/galleryService.js