## Summery
현재 개인 공간 중 사진 파편의 초기 생성 시간이 상당히 오래 걸리며, 그로 인해 블라킹 타임이 많이 발생하여 사용자의 경험을 해치는 문제가 있음.

## Context
개인 공간 중 사진 파편의 초기 생성 코드는 클린코드를 위해 스프레드 연산자를 이용해 각 루프마다 배열을 생성하고, 이를 분해하는 것을 반복하고 있었습니다. 하지만 이렇게 되면 사용하지 않은 오브젝트들이 많아져, 잦은 가비지 컬렉션이 발생하게 됩니다.
빠른 성능을 위해, 배열을 새로 선언하는 부분을 기존에 있던 큰 배열에 값을 푸시하는 방법으로 변경했습니다.

## Description
- 일부 상황에서 플레이어가 움직임을 지속하던 버그 수...정? (해치웠나?)
- 메인 페이지 FPS 향상(40->55FPS)
- 사진 파편 초기 생성 최적화(사진 파편 1개당 100ms -> 25ms)