@lybell-art 2022-12-12
와! Delayed를 걸었더니 blocking time이 더 떨어졌어요! (안 걸었을 때 100ms, 걸었을 때 1000ms)
Delayed는 안 거는 걸로 하겠습니다